Simplify dialog service
