--- web/src/app/components/confirmation-dialog/confirmation-dialog.component.ts
@@ -1,61 +1,0 @@
-/**
- * Copyright 2020 The Ground Authors.
- *
- * Licensed under the Apache License, Version 2.0 (the 'License');
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an 'AS IS' BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import {Component, Inject} from '@angular/core';
-import {MatButtonModule} from '@angular/material/button';
-import {
-  MAT_DIALOG_DATA,
-  MatDialogActions,
-  MatDialogClose,
-  MatDialogContent,
-  MatDialogRef,
-  MatDialogTitle,
-} from '@angular/material/dialog';
-
-@Component({
-  selector: 'ground-confirmation-dialog',
-  templateUrl: './confirmation-dialog.component.html',
-  standalone: true,
-  imports: [
-    MatDialogTitle,
-    MatDialogContent,
-    MatDialogActions,
-    MatDialogClose,
-    MatButtonModule,
-  ],
-})
-export class ConfirmationDialogComponent {
-  title: string;
-  message: string;
-  showDiscardActions = false;
-
-  constructor(
-    public dialogRef: MatDialogRef<ConfirmationDialogComponent>,
-    @Inject(MAT_DIALOG_DATA) public data: ConfirmationDialogComponent
-  ) {
-    this.title = data.title;
-    this.message = data.message;
-    this.showDiscardActions = data.showDiscardActions;
-  }
-
-  onConfirm() {
-    this.dialogRef.close(true);
-  }
-
-  onClose() {
-    this.dialogRef.close(false);
-  }
-}
